First proof of concept: doesn't work..
First try of loading a MAS where all the agents share the same OWL ontology.

It gives an error that, according to my "google" about it, is due to the ontology file sharing the same name for different things (i.e.: a property and a class sharing the same name). 

So, first, we need to have the OWL files refactored.